refactor canoe_verifier
