+++ crates/compute-proof/Cargo.toml
@@ -10,3 +10,4 @@ rust-kzg-bn254-primitives.workspace = true
 num.workspace = true
 
 alloy-primitives.workspace = true
+spin.workspace = true
