+++ example/canoe-on-sepolia/Cargo.toml
@@ -5,9 +5,11 @@ edition = "2021"
 
 [dependencies]
 tokio = { workspace = true, features = ["full"] }
-canoe-provider  = { workspace = true }
+canoe-provider = { workspace = true }
+canoe-verifier-address-fetcher = { workspace = true }
+canoe-verifier = { workspace = true }
 canoe-steel-apps = { workspace = true}
-hokulea-proof = { workspace = true, features = ["steel"] }
+canoe-steel-verifier = { workspace = true}
 eigenda-cert = { workspace = true }
 clap = { workspace = true, features = ["derive", "env"] }
 anyhow = { workspace = true }
